requirements.txt
# AI & APIs
anthropic>=0.18.0
elevenlabs>=0.3.0

# Vision & Screen
pillow>=10.0.0
mss>=9.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0

# Input Control
pyautogui>=0.9.54
pynput>=1.7.6

# YouTube Integration
google-api-python-client>=2.115.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pytchat>=0.5.0

# VTube Studio
websockets>=12.0
pyvts>=0.3.0

# Audio
pygame>=2.5.0
pydub>=0.25.1

# Data & Storage
aiosqlite>=0.19.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
aiohttp>=3.9.0
tenacity>=8.2.0
pyyaml>=6.0
